Replace deprecated in jdk8 JVM option
-XX:MaxPermSize=size Sets the maximum permanent generation space
size (in bytes).  This option was deprecated in JDK 8, and
superseded by the -XX:MaxMetaspaceSize option.

Before Java 8 class metadata space resided in PermGen which was
limited by 32/64MB, and MaxPerGen was used to increase it.
Starting from Java 8 however, PermGen is no more and class
metadata space is unlimited, so MaxMetaspace size is actually
used to decrease it.  For more information see:
stackoverflow.com/a/31463972/67824

i would set a maximum which is large enough to not trigger in
normal situations.  The reason I say this is, that a system might
act very erratic and hard to control if native memory gets
exhausted or wild swapping happens.  Much worse than a OOM or Java
freeze. For example using 3GB (epecting a system to have 3GB buffers
free at minimum).

See
https://stackoverflow.com/questions/22634644/java-hotspottm-64-bit-server-vm-warning-ignoring-option-maxpermsize
https://stackoverflow.com/questions/31455644/should-i-set-a-maxmetaspacesize